Cloud SQL/PgVector/requirements.txt
# asyncio==3.4.3 
# asyncpg==0.27.0 
# # pip install cloud-sql-python-connector["asyncpg"]==1.2.3
# numpy==1.22.4 
# pandas==1.5.3
# pgvector==0.1.8
# # langchain-google-vertexai
# # langchain==0.0.196 
# # google-cloud-aiplatform==1.26.0





asyncio==3.4.3
asyncpg==0.27.0
cloud-sql-python-connector==1.2.3
google-cloud-aiplatform==1.43.0
langchain==0.1.10
langchain-google-vertexai==0.1.0
numpy==1.22.4
pandas==1.5.3
pgvector==0.1.8




aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.7.0
cryptography==42.0.5
dataclasses-json==0.6.4
docarray==0.32.1
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.0
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
hnswlib==0.8.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.28
langchain-google-cloud-sql-pg==0.1.0
langchain-google-genai==0.0.9
langchain-google-vertexai==0.1.0
langchain-text-splitters==0.0.1
langsmith==0.1.13
markdown-it-py==3.0.0
marshmallow==3.21.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.9.15
packaging==23.2
pgvector==0.2.5
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==8.2.3
tqdm==4.66.2
types-protobuf==4.24.0.20240302
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
vertexai==0.0.1
yarl==1.9.4